+++ CHANGELOG.md
@@ -11,3 +11,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Command line interface parsing with [clap](https://github.com/clap-rs/clap).
 - 'new' command cli interface for starting a crame project.
+- Setup [tracing](https://github.com/tokio-rs/tracing) for application logging.